backend/requirements.txt
flask
flask-cors
flask-bcrypt
pymongo
tensorflow
pillow
numpy